Separate GzipAdnPutFilesWork into its own file
