+++ src/historywork/GzipAndPutFilesWork.cpp
@@ -1,5 +1,88 @@
-//
-// Created by Marta Lokhava on 10/29/18.
-//
+// Copyright 2018 Stellar Development Foundation and contributors. Licensed
+// under the Apache License, Version 2.0. See the COPYING file at the root
+// of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "GzipAndPutFilesWork.h"
+#include "bucket/BucketManager.h"
+#include "historywork/GzipFileWork.h"
+#include "historywork/MakeRemoteDirWork.h"
+#include "historywork/PutRemoteFileWork.h"
+
+namespace stellar
+{
+
+GzipAndPutFilesWork::GzipAndPutFilesWork(
+    Application& app, std::shared_ptr<HistoryArchive> archive,
+    std::shared_ptr<StateSnapshot> snapshot,
+    HistoryArchiveState const& remoteState)
+    : Work(app, "helper-put-files-" + archive->getName())
+    , mArchive(archive)
+    , mSnapshot(snapshot)
+    , mRemoteState(remoteState)
+{
+}
+
+BasicWork::State
+GzipAndPutFilesWork::doWork()
+{
+    if (!mChildrenSpawned)
+    {
+        std::vector<std::shared_ptr<FileTransferInfo>> files = {
+            mSnapshot->mLedgerSnapFile, mSnapshot->mTransactionSnapFile,
+            mSnapshot->mTransactionResultSnapFile,
+            mSnapshot->mSCPHistorySnapFile};
+
+        std::vector<std::string> bucketsToSend =
+            mSnapshot->mLocalState.differingBuckets(mRemoteState);
+
+        for (auto const& hash : bucketsToSend)
+        {
+            auto b = mApp.getBucketManager().getBucketByHash(hexToBin256(hash));
+            assert(b);
+            files.push_back(std::make_shared<FileTransferInfo>(*b));
+        }
+        for (auto f : files)
+        {
+            // If files are empty, they are removed and shouldn't be uploaded
+            if (f && fs::exists(f->localPath_nogz()))
+            {
+                auto gzipFile = std::make_shared<GzipFileWork>(
+                    mApp, f->localPath_nogz(), true);
+                auto mkdir = std::make_shared<MakeRemoteDirWork>(
+                    mApp, f->remoteDir(), mArchive);
+                auto putFile = std::make_shared<PutRemoteFileWork>(
+                    mApp, f->localPath_gz(), f->remoteName(), mArchive);
+
+                std::vector<std::shared_ptr<BasicWork>> seq{gzipFile, mkdir,
+                                                            putFile};
+
+                addWork<WorkSequence>("gzip-and-put-file-" + f->localPath_gz(),
+                                      seq);
+            }
+        }
+        mChildrenSpawned = true;
+    }
+    else
+    {
+        if (allChildrenSuccessful())
+        {
+            return State::WORK_SUCCESS;
+        }
+        else if (anyChildRaiseFailure())
+        {
+            return State::WORK_FAILURE;
+        }
+        else if (!anyChildRunning())
+        {
+            return State::WORK_WAITING;
+        }
+    }
+    return State::WORK_RUNNING;
+}
+
+void
+GzipAndPutFilesWork::doReset()
+{
+    mChildrenSpawned = false;
+}
+}
