+++ src/historywork/GzipAndPutFilesWork.h
@@ -1,14 +1,33 @@
-//
-// Created by Marta Lokhava on 10/29/18.
-//
+// Copyright 2018 Stellar Development Foundation and contributors. Licensed
+// under the Apache License, Version 2.0. See the COPYING file at the root
+// of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#ifndef STELLAR_CORE_GZIPANDPUTFILESWORK_H
-#define STELLAR_CORE_GZIPANDPUTFILESWORK_H
+#pragma once
 
+#include "history/FileTransferInfo.h"
+#include "history/HistoryArchive.h"
+#include "history/StateSnapshot.h"
+#include "work/Work.h"
 
-class GzipAndPutFilesWork {
+namespace stellar
+{
+class GzipAndPutFilesWork : public Work
+{
+    std::shared_ptr<HistoryArchive> mArchive;
+    std::shared_ptr<StateSnapshot> mSnapshot;
+    HistoryArchiveState const& mRemoteState;
 
+    bool mChildrenSpawned{false};
+
+  public:
+    GzipAndPutFilesWork(Application& app,
+                        std::shared_ptr<HistoryArchive> archive,
+                        std::shared_ptr<StateSnapshot> snapshot,
+                        HistoryArchiveState const& remoteState);
+    ~GzipAndPutFilesWork() = default;
+
+  protected:
+    void doReset() override;
+    State doWork() override;
 };
-
-
-#endif //STELLAR_CORE_GZIPANDPUTFILESWORK_H
+}
